config: SEGGER_RTT_Conf: Allow to use DT section
CONFIG_SEGGER_RTT_SECTION_CUSTOM_DTS_REGION was introduced some time
ago but it was not functional and memory section specified in the
DTS could not be used for RTT data. Kconfig string value cannot be
used for DTS section because a token without quotation marks is needed.

DT memory section need to have alias 'rtt_custom_section' and that
alias is used when CONFIG_SEGGER_RTT_SECTION_CUSTOM_DTS_REGION is
enabled.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>